fix: correct return type annotation in MinOptMax.__iter__()
Changed from hardcoded Iterator[float] to generic Iterator[Numeric]
to properly support both int and float types as defined by the class
generic parameter.

+++ prymer/api/minoptmax.py
@@ -71,7 +71,7 @@ class MinOptMax(Generic[Numeric]):
                 f"received min: {self.min}, opt: {self.opt}, max: {self.max}"
             )
 
-    def __iter__(self) -> Iterator[float]:
+    def __iter__(self) -> Iterator[Numeric]:
         """Returns an iterator of min, opt, and max"""
         return iter([self.min, self.opt, self.max])
